split widget bases into files
